+++ components/organisms/Header/header.tsx
@@ -5,7 +5,7 @@ import Title from "../../atoms/Typography/title";
 import ContextThumbnail from "../../atoms/ContextThumbnail/context-thumbnail";
 import FilterCard from "../../atoms/FilterCard/filter-card";
 import Thumbnail from "../../../public/hacktoberfest-icon.png";
-import PillSelector from "components/atoms/PillSelector/pill-selector";
+import SuperativeSelector from "components/molecules/SuperlativeSelector/superlative-selector";
 import useFilterOptions from "lib/hooks/useFilterOptions";
 
 const Header: React.FC = () => {
@@ -25,17 +25,26 @@ const Header: React.FC = () => {
   return (
     <section className="header flex flex-col md:flex-row pt-6 px-4 md:px-16 bg-slate-50">
       <div className="header-image mr-2 p-2 min-w-[130px]">
-        <ContextThumbnail 
-          size={120} ContextThumbnailURL={Thumbnail.src}></ContextThumbnail>
+        <ContextThumbnail size={120} ContextThumbnailURL={Thumbnail.src}></ContextThumbnail>
       </div>
       <div className="header-info flex flex-col grow justify-center p-2">
-        <Title level={1} className="!text-3xl font-semibold tracking-tight text-slate-900">Hacktoberfest 2022</Title>
-        <Text className="mt-1 !text-base font-medium text-slate-500">Insights on GitHub project opted into the largest open source hackathon.</Text>
-        <div className="flex mt-4 items-center gap-2">
+        <Title level={1} className="!text-3xl font-semibold tracking-tight text-slate-900">
+          Hacktoberfest 2022
+        </Title>
+        <Text className="mt-1 !text-base font-medium text-slate-500">
+          Insights on GitHub project opted into the largest open source hackathon.
+        </Text>
+        <div className="flex mt-4 items-center relative  gap-2">
           <FilterCard filterName="hacktoberfest" isRemovable={false} icon="topic" />
-          <PillSelector
+          {/* <PillSelector
             pillOptions={filterOptions}
             handlePillClick={filterBtnRouting}
+            handleCancelClick={cancelFilterRouting}
+            selected={selectedFilter as string}
+          /> */}
+          <SuperativeSelector
+            filterOptions={filterOptions}
+            handleFilterClick={filterBtnRouting}
             handleCancelClick={cancelFilterRouting}
             selected={selectedFilter as string}
           />