+++ stories/organisms/dashboard.stories.tsx
@@ -1,0 +1,9 @@
+import Dashboard from "components/organisms/Dashboard/dashboard";
+
+
+const StoryConfig = {
+  title: "Design System/organisms/Dashboard"
+};
+export default StoryConfig;
+
+export const DashboardOrganism = (): JSX.Element => <Dashboard />;