+++ stories/organisms/repository.stories.tsx
@@ -1,0 +1,9 @@
+import Repositories from "components/organisms/Repositories/repositories";
+
+const StoryConfig = {
+  title: "Design System/organisms/Repositories"
+};
+
+export default StoryConfig;
+
+export const ReportsOrganism = (): JSX.Element => <Repositories />;